requirements.txt
click
jinja2
pyyaml
pytest
coverage
zerotk.easyfs
